# Conflicts:
#	core/src/main/java/com/solar/MainGame.java
#	core/src/main/java/com/solar/data/PlanetData.java
#	core/src/main/java/com/solar/data/PlanetDatabase.java
#	core/src/main/java/com/solar/screen/BaseScreen.java
#	core/src/main/java/com/solar/screen/GameScreen.java
#	core/src/main/java/com/solar/screen/LoadingScreen.java
#	core/src/main/java/com/solar/screen/MenuScreen.java
#	core/src/main/java/com/solar/screen/PlanetScreen.java
#	core/src/main/java/com/solar/screen/SolarSystemScreen.java
#	core/src/main/java/com/solar/ui/GameHud.java

+++ core/src/main/java/com/solar/data/PlanetData.java
@@ -1,26 +1,33 @@
 package com.solar.data;
 
+// Stores all data for a planet
 public class PlanetData {
-
+    // Planet type (enum)
     public PlanetType type;
+    // Display name for UI
     public String displayName;
+    // Texture region name in atlas
     public String texturePath;
-    public float size;
-    public boolean canEnter;
-    // Vị trí lưu tạm cho Moon
+    public float size; // Size of the planet (pixels)
+    public boolean canEnter;     // Can player enter this planet?
+
+    // Temporary position for Moon
     public float positionX;
+    // X position in solar system
     public float x;
+    // Y position in solar system
     public float y;
 
-    public float gravity;       // Trọng lực
-    public String weather;  // Nhiệt độ
-    public String atmosphere;   // Khí quyển
-    public String surfaceType;  // Bề mặt
-    public String primaryRes;   // Tài nguyên
+    // Physics and info
+    public double gravity;       // Gravity value
+    public String weather;       // Weather description
+    public String atmosphere;    // Atmosphere type
+    public String surfaceType;   // Surface type
+    public String primaryRes;    // Main resource
 
-
+    // Constructor: initializes all planet data
     public PlanetData(PlanetType type, String displayName, String texturePath, float size, boolean canEnter, float x, float y,
-                      float gravity, String weather, String atmosphere, String surfaceType, String primaryRes) {
+                      double gravity, String weather, String atmosphere, String surfaceType, String primaryRes) {
         this.type = type;
         this.displayName = displayName;
         this.texturePath = texturePath;
@@ -29,12 +36,10 @@ public class PlanetData {
         this.positionX = 0;
         this.x = x;
         this.y = y;
-
         this.gravity = gravity;
         this.weather = weather;
         this.atmosphere = atmosphere;
         this.surfaceType = surfaceType;
         this.primaryRes = primaryRes;
-
     }
 }
